% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stge-package.R
\docType{package}
\name{stge-package}
\alias{stge}
\alias{stge-package}
\title{stge: Estimation for Spatio-Temporal Gene Expression pattern}
\description{
This package is inteded to estimate spatio-temporal expression patters by combining cell migration data and single cell RNA seq data.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
